# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kuras Pricer public API
  version: 1.0.0
  description: >-
    Versioned read-only API for official daily LEA fuel prices. LEA-derived
    prices are published on working days and must not be described as real-time.
servers:
  - url: https://kuras.pricer.lt/api/v1
paths:
...
  /rankings:
    get:
      summary: Cheapest stations by scope and fuel
      operationId: getRankings
      parameters:
        - { $ref: '#/components/parameters/Fuel' }
        - { $ref: '#/components/parameters/Date' }
        - { name: municipality, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
      responses:
        '200': { $ref: '#/components/responses/CollectionResponse' }
  /statistics:
    get:
      summary: National and municipality price aggregates
      operationId: getStatistics
      parameters:
        - { $ref: '#/components/parameters/Fuel' }
        - { $ref: '#/components/parameters/Date' }
      responses:
        '200': { $ref: '#/components/responses/SingleResponse' }
components:
  parameters:
    Fuel:
      name: fuel
      in: query
      schema: { type: string, enum: [pb95, pb98, diesel, lpg], default: pb95 }
    Date:
      name: date
      in: query
      description: Defaults to the last successfully published source date.
      schema: { type: string, format: date }
...
  schemas:
...
    SourceMeta:
      type: object
      required: [api_version, source_date, source, freshness]
      properties:
        api_version: { type: string, const: v1 }
        source_date: { type: string, format: date }
        last_successful_import_at: { type: [string, 'null'] }
        source: { type: object }
        freshness: { type: object }
...
  responses:
    CollectionResponse:
      description: Collection and source metadata
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data: { type: array, items: {} }
              meta: { $ref: '#/components/schemas/SourceMeta' }
    SingleResponse:
      description: Resource and source metadata
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data: { type: object }
              meta: { $ref: '#/components/schemas/SourceMeta' }
...